+++ pygeotemporal_parsers/parse/get_convert_concat_parse_upload.py
@@ -2,13 +2,13 @@
     * Parse and Upload the Newest Locally-Present Data File *
 
     This Script will do the following:
-        - Get the files for parsing 
+        - Get the files for parsing
+        - Convert XLS files to CSV files
         - Concat the new files and create a parsing file
         - Create Sensors and Streams if they do not exist
         - Parse the new Data
         - Update the Sensors Stats
         - Upload newest Aggregate to the correct location
-
 """
 
 # Python Imports
@@ -30,6 +30,7 @@ from parse_new_datapoints import parse_data, update_sensors_stats
 from get_new_files import filter_files
 from concat_files import create_header, concat_files
 from parse_and_upload_newest_file import upload_file
+from convert_xls_to_csv import convert_files
 
 
 def main():
@@ -113,6 +114,14 @@ def main():
 
     # GET NEW FILES end #
 
+    # CONVERT FILES start #
+
+    # Convert File
+    print("Converting XLS Files to CSV Files. ")
+    convert_files(local_path, new_files, downloads)
+
+    # CONVERT FILES end #
+
     # CONCAT FILES start #
 
     print "INFO: file to write is: " + str(concat_file_name)
@@ -193,7 +202,6 @@ def main():
     # UPLOAD NEW AGGREGATE FILE end #
 
     print("Processing Complete. ")
-
 
 if __name__ == '__main__':
     try: