+++ pygeotemporal_parsers/parse/get_convert_concat_rename_parse_upload.py
@@ -2,13 +2,14 @@
     * Parse and Upload the Newest Locally-Present Data File *
 
     This Script will do the following:
-        - Get the files for parsing 
+        - Get the files for parsing
+        - Convert XLS files to CSV files
         - Concat the new files and create a parsing file
         - Create Sensors and Streams if they do not exist
         - Parse the new Data
+        - Use New Parameter Names
         - Update the Sensors Stats
         - Upload newest Aggregate to the correct location
-
 """
 
 # Python Imports
@@ -26,10 +27,11 @@ from pyclowder.datasets import get_file_list
 # Module Imports
 from pygeotemporal_parsers.sensors.create_sensors_and_streams \
     import create_sensors_and_streams
-from parse_new_datapoints import parse_data, update_sensors_stats
+from parse_new_datapoints_new_params import parse_data, update_sensors_stats
 from get_new_files import filter_files
 from concat_files import create_header, concat_files
 from parse_and_upload_newest_file import upload_file
+from convert_xls_to_csv import convert_files
 
 
 def main():
@@ -67,6 +69,7 @@ def main():
     verify_header = multi_config['inputs']['verify']
     timestamp = multi_config['inputs']['timestamp']
     parameters = multi_config['parameters']
+    parameters_updated = multi_config['parameters_updated']
     sensor_names = multi_config['sensors']
     total_header_rows = int(multi_config['inputs']['headers'])
     sensor_names_create = multi_config['sensorscreate']
@@ -112,6 +115,14 @@ def main():
     output_file.close()
 
     # GET NEW FILES end #
+
+    # CONVERT FILES start #
+
+    # Convert File
+    print("Converting XLS Files to CSV Files. ")
+    convert_files(local_path, new_files, downloads)
+
+    # CONVERT FILES end #
 
     # CONCAT FILES start #
 
@@ -167,8 +178,8 @@ def main():
     # Parse Data
     print("Will parse data. ")
     parse_file = open(parse_file_name, 'r')
-    parse_data(timestamp, config, sensor_names, parameters, parse_file,
-               sensor_client, stream_client, datapoint_client)
+    parse_data(timestamp, config, sensor_names, parameters, parameters_updated,
+               parse_file, sensor_client, stream_client, datapoint_client)
 
     # Update the sensors
     print("Will update sensor stats. ")
@@ -193,7 +204,6 @@ def main():
     # UPLOAD NEW AGGREGATE FILE end #
 
     print("Processing Complete. ")
-
 
 if __name__ == '__main__':
     try: